Use u64 indices in TreeListModel.
The GTK ListModel API currently only supports 32-bit row numbers, but we
don't want to be limited by this.

Handle this by having the GObject wrapper restrict what's exposed to GTK
to the first 2^32-1 rows, whilst keeping full 64-bit indices internally
in TreeListModel, so that the limitation can be lifted easily in future.